Visual brief
These images represent what we're going for. The flat-bottom taco is what we want playfully illustrated. We also want the taco to bend out slightly to look like a "V". We don't have to represent the "V" exactly, just slightly. We're more interested in the illustration looking like a taco. As for the colors, the taco should be yellow with darker outlines. The background should be purple as this is how it will be represented in our marketing.